+++ NEWS.md
@@ -1,3 +1,8 @@
+# scico 1.3.0
+
+* Added new colour palettes (@michaelgrund, #11)
+* Added midpoint argument to ggplot2 scales
+
 # scico 1.2.0
 
 * Added cyclic colour palettes